adding update function to cleanup

+++ src/cleanup.py
@@ -27,7 +27,7 @@ import shutil
 import glob
 import signal
 import subprocess
-from utils import kill_process, Dict
+from utils import kill_process, Dict, process_create_time
 
 
 def retrieve_catalog(cfg):
@@ -96,13 +96,51 @@ def local_rmdir(cfg, dirname):
 def cancel_job(job_num,qsys):
         if job_num is not None:
             logging.info('Deleting parallel job %s on %s.' % (job_num, qsys))
-            clusters = json.load(open('etc/clusters.json','r'))
-            cluster = clusters[qsys]
+            cluster = load_cluster_file(qsys)
             try:
                 ret = subprocess.check_output([cluster['qdel_cmd'], job_num])
                 logging.info(ret)
             except:
                 logging.error('Deleting parallel job %s failed.' % job_num)
+
+def load_cluster_file(qsys):
+    clusters = json.load(open('etc/clusters.json','r'))
+    return clusters[qsys]
+
+def load_job_file(job_id,cfg):
+    jobfile = osp.join(cfg['workspace_path'], job_id, 'job.json')
+    logging.info('Loading job state from %s' % jobfile)
+    try:
+        js = Dict(json.load(open(jobfile,'r')))
+    except IOError:
+        logging.error('Cannot open %s' % jobfile)
+        js = None
+    except:
+        logging.error('Cannot load %s' % jobfile)
+        js = None
+    # print(json.dumps(js, indent=4, separators=(',', ': ')))
+    return js, jobfile
+
+def forecast_process_running(js):
+    if 'pid' in js and 'process_create_time' in js:
+        fs =  js.process_create_time == process_create_time(js.pid)
+        logging.info('Forecast process is running: %s pid=%s' % (fs, js.pid))
+        return fs
+    else:
+        logging.warning('Cannot determine if forecast process is running - old job file?')
+        return None 
+
+def parallel_job_running(js):
+    cluster = load_cluster_file(js.qsys)
+    ret = subprocess.check_output(cluster['qstat_cmd'])
+    ws = False
+    for line in ret.split('\n'):
+        ls=line.split()
+        if len(ls) >0 and ls[0] == js.job_num:
+             ws = True
+             break
+    logging.info('WRF job is running: %s name=%s' % (ws, js.job_num))
+    return ws
 
 # the cleanup functions called as argv[2]:
 
@@ -142,23 +180,14 @@ def output(name,cfg):
 
 def cancel(name,cfg):
     logging.info('Trying to cancel job %s' % name)
-    jobfile = osp.join(cfg['workspace_path'], name, 'job.json')
-    logging.info('Loading job state from %s' % jobfile)
-    try:
-        js = Dict(json.load(open(jobfile,'r')))
-    except IOError:
-        logging.error('Cannot open %s' % jobfile)
-    except:
-        logging.error('Cannot load %s' % jobfile)
-    else:
-        # print(json.dumps(js, indent=4, separators=(',', ': ')))
+    js, jobfile = load_job_file(name,cfg)
+    if js is not None:
         if 'job_num' in js:
             cancel_job(js.job_num,js.qsys)
             js.old_job_num = js.job_num
         js.job_num = None
         if 'pid' in js:
             kill_process(js.pid)
-            js.old_pid = js.pid
         js.pid = None
         js.state = "Cancelled"
         json.dump(js, open(jobfile,'w'), indent=4, separators=(',', ': '))
@@ -175,11 +204,46 @@ def delete(name,cfg):
             del cat[name]
             store_catalog(cfg, cat)
 
+def update(name,cfg):
+    logging.info('Updating state of job %s' % name)
+    js, jobfile = load_job_file(name,cfg)
+    if js is not None:
+        logging.info('Old state is: %s' % js.state)
+        if not forecast_process_running(js):
+            #logging.info('Forecast process is not running')
+            js.pid = None
+        if not parallel_job_running(js):    
+            #logging.info('WRF is not running')
+            js.old_job_num = js.job_num
+            js.job_num = None
+        if js.state == 'Completed' or js.state == 'Cancelled':
+            if js.pid is not None:
+                 js.state = 'Forecast runaway'
+            if js.job_num is not None:
+                 js.state = 'WRF runaway'
+        elif js.state == 'Preparing':
+                 if js.pid is None:
+                      js.state = 'Stopped'
+                 if js.job_num is not None:
+                      js.state = 'Cannot happen'
+        elif js.state == 'Processing':
+                 if js.pid is None and js.job_num is None:
+                      js.state = 'Stopped'
+                 if js.pid is not None and js.job_num is None:
+                      js.state = 'WRF stopped'
+                 if js.pid is None and js.job_num is not None:
+                      js.state = 'Forecast stopped'
+
+        logging.info('New state is: %s' % js.state)
+
+
+        json.dump(js, open(jobfile,'w'), indent=4, separators=(',', ': '))
+
 if __name__ == '__main__':
 
     logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 
-    commands = [ 'list', 'cancel', 'output', 'delete', 'workspace']
+    commands = [ 'list', 'cancel', 'output', 'delete', 'workspace', 'update']
 
     if len(sys.argv) < 2 or sys.argv[1] not in commands: 
         print('usage: ./cleanup.sh ' + '|'.join(commands) +' [job_id]')
@@ -188,10 +252,11 @@ if __name__ == '__main__':
         print('output <job id> : cancel, and delete all WRF output and visualization files only')
         print('delete <job_id> : cancel, and delete all files')
         print('workspace       : delete jobs that are not on the visulalization server')
+        print('update <job_id> : check if the job is running and update its job state file')
         sys.exit(1)
 
     cmd = sys.argv[1]
-    if cmd in ['delete' , 'cancel', 'output']: 
+    if cmd in ['delete' , 'cancel', 'output', 'update']: 
         if len(sys.argv) == 3 and not sys.argv[2] == "" :
             job_id = sys.argv[2]
         else:
@@ -220,3 +285,6 @@ if __name__ == '__main__':
 
     if cmd == 'workspace':
         workspace(cfg)
+
+    if cmd == 'update':
+        update(job_id,cfg)
